:sparkles: Simplify API & Fix IDL generation
