+++ crates/bolt-lang/Cargo.toml
@@ -9,8 +9,8 @@ license = "MIT"
 anchor-lang = { version = "0.29.0"}
 
 # Bolt Attributes
+bolt-attribute-bolt-program = { path = "./attribute/bolt-program", version = "0.0.1" }
 bolt-attribute-bolt-component = { path = "./attribute/component", version = "0.0.1" }
-bolt-attribute-bolt-account = { path = "./attribute/account", version = "0.0.1" }
 bolt-attribute-bolt-system = { path = "./attribute/system", version = "0.0.1" }
 bolt-attribute-bolt-component-deserialize = { path = "./attribute/component-deserialize", version = "0.0.1" }
