+++ examples/system-apply-velocity/src/lib.rs
@@ -5,10 +5,7 @@ use component_velocity::Velocity;
 declare_id!("6LHhFVwif6N9Po3jHtSmMVtPjF6zRfL3xMosSzcrQAS8");
 
 #[system]
-#[program]
 pub mod system_apply_velocity {
-    use super::*;
-
     pub fn execute(ctx: Context<Component>, _args: Vec<u8>) -> Result<(Velocity, Position)> {
         ctx.accounts.velocity.x = 10;
         let clock = Clock::get()?;
