+++ programs/bolt-component/src/lib.rs
@@ -6,7 +6,14 @@ declare_id!("CmP2djJgABZ4cRokm4ndxuq6LerqpNHLBsaUv2XKEJua");
 pub mod bolt_component {
     use super::*;
 
-    pub fn initialize(_ctx: Context<Initialize>) -> Result<()> {
+    pub fn initialize(ctx: Context<Initialize>) -> Result<()> {
+        // ctx.accounts.data.to_account_info().assign(::ID); TODO: Test delegate to the world program
+        if let Some(authority) = &ctx.accounts.authority {
+            if authority.key != ctx.accounts.payer.key {
+                panic!("Authority mismatch");
+            }
+            ctx.accounts.data.bolt_metadata.authority = *authority.key;
+        }
         Ok(())
     }
 
@@ -17,7 +24,7 @@ pub mod bolt_component {
     #[derive(Accounts)]
     pub struct Apply<'info> {
         #[account(mut)]
-        pub bolt_component: Account<'info, ComponentData>,
+        pub bolt_component: Account<'info, Component>,
         /// CHECK: The system can modify the data of the component
         pub bolt_system: UncheckedAccount<'info>,
     }
@@ -41,7 +48,7 @@ pub mod bolt_component {
     #[derive(Accounts)]
     pub struct Update<'info> {
         #[account(mut)]
-        pub bolt_component: Account<'info, ComponentData>,
+        pub bolt_component: Account<'info, Component>,
     }
 }
 
@@ -49,25 +56,28 @@ pub mod bolt_component {
 pub struct Initialize<'info> {
     #[account(mut)]
     pub payer: Signer<'info>,
-    #[account(init_if_needed, payer = payer, space = ComponentData::size(), seeds = [ComponentData::seed(), entity.key().as_ref()], bump)]
-    pub data: Account<'info, ComponentData>,
+    #[account(init_if_needed, payer = payer, space = Component::size(), seeds = [Component::seed(), entity.key().as_ref()], bump)]
+    pub data: Account<'info, Component>,
     #[account()]
     /// CHECK: A generic entity account
-    pub entity: UncheckedAccount<'info>,
+    pub entity: AccountInfo<'info>,
+    #[account()]
+    /// CHECK: The authority of the component
+    pub authority: Option<AccountInfo<'info>>,
     pub system_program: Program<'info, System>,
 }
 
 // Component data
 #[account]
 #[derive(InitSpace, Default, Copy)]
-pub struct ComponentData {
-    pub id: u64,
+pub struct Component {
     pub position: Position,
+    pub bolt_metadata: BoltMetadata,
 }
 
-impl ComponentData {
+impl Component {
     pub fn size() -> usize {
-        8 + ComponentData::INIT_SPACE
+        8 + Component::INIT_SPACE
     }
     pub fn seed() -> &'static [u8] {
         b"origin-component"
@@ -79,4 +89,9 @@ pub struct Position {
     pub x: i64,
     pub y: i64,
     pub z: i64,
+}
+
+#[derive(InitSpace, AnchorSerialize, AnchorDeserialize, Default, Copy, Clone)]
+pub struct BoltMetadata {
+    pub authority: Pubkey
 }
